requirements.txt
streamlit
pandas
pillow
ultralytics
opencv-python
twilio